update modle structure
